Move useful stuff to common
