+++ packages/dataproviders/src/util/index.ts
@@ -1,8 +1,8 @@
 export * from './codec';
 export * from './constants';
+export * from './countries';
 export * from './errors';
 export * from './fp';
-export * from './getCountryCode';
 export * from './getDominantPol';
 export * from './openaq';
 export * from './stationName';
