+++ packages/ui/package.json
@@ -21,8 +21,10 @@
 	"dependencies": {
 		"@shootismoke/dataproviders": "^0.7.0",
 		"@sindresorhus/slugify": "^1.1.0",
+		"@types/async-retry": "^1.4.2",
 		"@types/haversine": "^1.1.4",
 		"@types/p-any": "^1.1.3",
+		"async-retry": "^1.3.1",
 		"expo-server-sdk": "^3.6.0",
 		"haversine": "^1.1.1",
 		"lottie-react-native": "^3.4.0",
